refactor: TabsLayout 리팩토링 진행
