--- fourtoon-cookie/src/pages/CharacterSelectPage/TabsLayout/Tab/Tab.styled.ts
@@ -1,32 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({
-    tab: {
-        flexDirection: 'row',
-        justifyContent: 'center',
-        alignItems: 'center',
-        paddingVertical: 8,
-        paddingHorizontal: 12,
-        borderRadius: 18,
-        marginRight: 8,
-    },
-    selectedTab: {
-        backgroundColor: '#212121',
-    },
-    tabText: {
-        fontFamily: 'Pretendard',
-        fontWeight: '500',
-        fontSize: 16,
-        lineHeight: 20,
-        letterSpacing: -0.005,
-        color: '#AAAAAA',
-    },
-    activeTabText: {
-        fontFamily: 'Pretendard',
-        fontWeight: '700',
-        fontSize: 16,
-        lineHeight: 20,
-        letterSpacing: -0.005,
-        color: '#FFFFFF',
-    }
-});
