--- fourtoon-cookie/src/pages/CharacterSelectPage/TabsLayout/Tab/Tab.tsx
@@ -1,21 +1,0 @@
-import React from 'react';
-import { Text, TouchableOpacity } from 'react-native';
-import * as S from './Tab.styled';
-
-export interface TabProps {
-    isActive: boolean;
-    label: string;
-    onPress: () => void;
-}
-
-const Tab = (props: TabProps) => {
-    const { isActive, label, onPress, ...rest } = props;
-
-    return (
-        <TouchableOpacity onPress={onPress} style={[S.styles.tab, isActive && S.styles.selectedTab ]}>
-            <Text style={[S.styles.tabText, isActive && S.styles.activeTabText]}>{label}</Text>
-        </TouchableOpacity>
-    );
-};
-
-export default Tab;
