+++ fourtoon-cookie/src/pages/CharacterSelectPage/TabsLayout/TabsLayout.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import { View } from 'react-native';
-import Tab from './Tab/Tab';
+import { Text, TouchableOpacity, View } from 'react-native';
 import * as S from './TabsLayout.styled';
 import { FREE_KOR, PAID_KOR } from '../../../constants/character';
 import { CharacterPaymentType } from '../../../types/character';
@@ -15,18 +14,36 @@ const TabsLayout = (props: TabsLayoutProps) => {
     
     return (
         <View style={S.styles.headersContainer}>
-            <Tab 
-                isActive={selectedPaymentType === CharacterPaymentType.FREE} 
-                label={FREE_KOR}
-                onPress={() => onSelectedPaymentTypeChange(CharacterPaymentType.FREE)} 
-            />
-            {/* <Tab 
-                isActive={selectedPaymentType === CharacterPaymentType.PAID} 
-                label={PAID_KOR}
-                onPress={() => onSelectedPaymentTypeChange(CharacterPaymentType.PAID)} 
-            /> */}
+            {
+                [
+                    { type: CharacterPaymentType.FREE, label: FREE_KOR},
+                    //{ type: CharacterPaymentType.PAID, label: PAID_KOR}
+                ].map((item) => (
+                    <Tab
+                        isActive={selectedPaymentType === item.type}
+                        label={item.label}
+                        onPress={() => onSelectedPaymentTypeChange(item.type)}
+                    />
+                ))
+            }
         </View>
     );
 };
 
 export default TabsLayout;
+
+interface TabProps {
+    isActive: boolean;
+    label: string;
+    onPress: () => void;
+}
+
+const Tab = (props: TabProps) => {
+    const { isActive, label, onPress, ...rest } = props;
+
+    return (
+        <TouchableOpacity onPress={onPress} style={[S.styles.tab, isActive && S.styles.selectedTab ]}>
+            <Text style={[S.styles.tabText, isActive && S.styles.activeTabText]}>{label}</Text>
+        </TouchableOpacity>
+    );
+};
